requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.14.0
django-filter>=23.2
drf-yasg>=1.21.7
pytest>=7.3.1
pytest-django>=4.5.2
pytest-cov>=4.1.0
bandit>=1.7.5
flake8>=6.0.0
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
django-stubs>=4.2.0
djangorestframework-stubs>=3.14.0
